Create Gaussian Quadrature
The program provided is gaussint.py. It includes
a ‘library’ function gaussint(a,b,delta) that uses Gauss’ method. The arguments
a and b are the range of integration and delta specifies the accuracy you want; the
range is subdivided until the relative change in the result is less than delta. With an
easy integral you should be able to get 1 part in 106 but with a difficult integral you
may have to be satisfied with less accuracy. You have to specify the function you are
integrating and this is done in the function f(x). The programs include a test function
and an example of how to use the library